requirements.txt
sqlalchemy ~= 1.3.20
pandas ~= 1.1.4
python-dotenv ~= 0.15.0
pyodbc ~= 4.0.30
colorama ~= 0.4.4
PyYAML~=5.3.1
sqlacodegen~=2.3.0